pyproject.toml
[project]
name = "app-size-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=2.18.2",
    "pandas>=2.2.3",
    "pyelftools>=0.32",
]
